:recycle: Refactor Estructura
