+++ src/main/java/com/alumnione/ecommerce/dto/CartCreationDto.java
@@ -1,4 +1,4 @@
-package com.alumnione.ecommerce.model.cart.dto;
+package com.alumnione.ecommerce.dto;
 
 
 public record CartCreationDto (){} //TODO: Agregar campos necesarios para la creación de un carrito (lo tenemos que definir mas adelante)
