+++ src/main/java/com/alumnione/ecommerce/repository/CartRepository.java
@@ -1,8 +1,8 @@
-package com.alumnione.ecommerce.model.cart.repository;
+package com.alumnione.ecommerce.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.alumnione.ecommerce.model.cart.entity.Cart;
+import com.alumnione.ecommerce.entity.Cart;
 
 public interface CartRepository extends JpaRepository<Cart, Long> { 
 }
