+++ src/main/java/com/alumnione/ecommerce/repository/CellphoneRepository.java
@@ -1,9 +1,9 @@
-package com.alumnione.ecommerce.model.cellphone.repository;
+package com.alumnione.ecommerce.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.alumnione.ecommerce.model.cellphone.entity.Cellphone;
+import com.alumnione.ecommerce.entity.Cellphone;
 
 
 @Repository
