+++ src/main/java/com/alumnione/ecommerce/repository/InvoiceRepository.java
@@ -1,6 +1,6 @@
-package com.alumnione.ecommerce.model.invoice.repository;
+package com.alumnione.ecommerce.repository;
 
-import com.alumnione.ecommerce.model.invoice.entity.Invoice;
+import com.alumnione.ecommerce.entity.Invoice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
