+++ src/main/java/com/alumnione/ecommerce/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package com.alumnione.ecommerce.model.payment.repository;
+package com.alumnione.ecommerce.repository;
 
-import com.alumnione.ecommerce.model.payment.entity.Payment;
+import com.alumnione.ecommerce.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
